Move store location into core
