+++ e-online-shop/src/app/app.config.ts
@@ -3,12 +3,12 @@ import { provideRouter, withComponentInputBinding } from '@angular/router';
 import { routes } from './app.routes';
 import { provideClientHydration } from '@angular/platform-browser';
 import { provideStore } from '@ngrx/store';
-import { productReducer } from './store/products/products.reducer';
+import { productReducer } from './core/store/products/products.reducer';
 import { provideStoreDevtools } from '@ngrx/store-devtools';
 import { provideEffects } from '@ngrx/effects';
-import { ProductEffects } from './store/products/products.effects';
-import { CartEffects } from './store/cart/cart.effects';
-import { cartReducer } from './store/cart/cart.reducer';
+import { ProductEffects } from './core/store/products/products.effects';
+import { CartEffects } from './core/store/cart/cart.effects';
+import { cartReducer } from './core/store/cart/cart.reducer';
 
 export const appConfig: ApplicationConfig = {
 	providers: [
