+++ e-online-shop/src/app/core/store/cart/cart.actions.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { Product } from '../../shared/models/product.model';
+import { Product } from '../../../shared/models/product.model';
 
 export const addToCart = createAction(
 	'[Cart] Add to Cart',
