+++ e-online-shop/src/app/core/store/products/products.reducer.ts
@@ -1,6 +1,6 @@
 import { createReducer, on } from '@ngrx/store';
 import { loadProductsSuccess } from './products.actions';
-import { Product } from '../../shared/models/product.model';
+import { Product } from '../../../shared/models/product.model';
 
 export interface ProductsState {
 	products: Product[];
